rabbitmq bosh cats tests
